+++ gb-bus/src/error.rs
